Fixed plenty bugs and realized so many packets are buggy
-This version still has many bugged packets

+++ src/main/java/packets/incoming/ReconnectPacket.java
@@ -42,7 +42,7 @@ public class ReconnectPacket extends Packet {
         host = buffer.readString();
         port = buffer.readInt();
         gameId = buffer.readInt();
-        keyTime = buffer.readInt();
+        keyTime = buffer.readByte();
         isFromArena = buffer.readBoolean();
         key = buffer.readByteArray();
     }
